requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for inter-service communication
httpx==0.25.2

# OpenTelemetry core
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# OpenTelemetry instrumentation
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0

# OpenTelemetry exporters
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0

# Structured logging
structlog==23.2.0

# Data validation and parsing
pydantic==2.5.0

# Environment management
python-dotenv==1.0.0